Use int32 for Generation fields and implement NodePools support
Changes:
- Change internal types from int64 to int32 for Generation and ObservedGeneration
  to match OpenAPI spec v1.0.0 exactly
- Implement full support for NodePools endpoint (GET /api/hyperfleet/v1/nodepools)
- Implement search parameter support for label selector filtering
- Add labelSelectorToSearchString() helper with alphabetical sorting
- Refactor fetchResourcesOnce() to use switch statement for resource types
- Split fetch logic into fetchClusters() and fetchNodePools()

Tests:
- Add TestLabelSelectorToSearchString with 5 test cases
- Add TestFetchResources_NodePools to verify nodepool fetching
- Add TestFetchResources_WithLabelSelector to verify search parameter
- Update decision_test.go helper functions to use int32

Removed:
- Outdated TODOs about resource types (spec v1.0.0 has nodepools)
- Outdated TODOs about search parameter (spec v1.0.0 has SearchParams)
- Outdated TODOs about int32 consideration (now using int32 directly)

All tests passing with 76.8% coverage in client package.

+++ internal/engine/decision_test.go
@@ -42,7 +42,7 @@ func newTestResourceWithCreatedTime(id, kind, phase string, createdTime, lastUpd
 }
 
 // newTestResourceWithGeneration creates a test resource with explicit generation values
-func newTestResourceWithGeneration(id, kind, phase string, lastUpdated time.Time, generation, observedGeneration int64) *client.Resource {
+func newTestResourceWithGeneration(id, kind, phase string, lastUpdated time.Time, generation, observedGeneration int32) *client.Resource {
 	return &client.Resource{
 		ID:          id,
 		Kind:        kind,
@@ -590,8 +590,8 @@ func TestDecisionEngine_Evaluate_GenerationBasedReconciliation(t *testing.T) {
 
 	tests := []struct {
 		name               string
-		generation         int64
-		observedGeneration int64
+		generation         int32
+		observedGeneration int32
 		phase              string
 		lastUpdated        time.Time
 		wantShouldPublish  bool
